Move ImageLoaderFactoryInitializer functionality into ChatUIInitializer
